+++ packages/ng-diagram/projects/ng-diagram/src/core/src/edge-routing-manager/index.ts
@@ -1,5 +1,6 @@
 export * from './edge-routing-manager';
 export { BUILT_IN_EDGE_ROUTINGS } from './edge-routing-manager';
+export * from './resolve-label-position';
 export * from './routings/bezier/bezier-routing';
 export * from './routings/orthogonal/orthogonal-routing';
 export * from './routings/polyline/polyline-routing';